% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hommap.R
\name{generateDataObjects}
\alias{generateDataObjects}
\title{Used during development to generate data objects that are saved with this package}
\usage{
generateDataObjects()
}
\description{
Used during development to generate data objects that are saved with this package
}
