% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hommap.R
\name{hommap.mm.2.hs.symbols}
\alias{hommap.mm.2.hs.symbols}
\title{Get symbol level homology map between human and mouse}
\usage{
hommap.mm.2.hs.symbols()
}
\description{
Get symbol level homology map between human and mouse
}
