% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs.mm.homologue-data.R
\docType{data(hs.mm.homologue)}
\name{hs.mm.homologue}
\alias{hs.mm.homologue}
\title{Mouse to human gene mapping from ensembl}
\format{An object of class \code{'tibble'}}
\usage{
hs.mm.homologue
}
\description{
Data generated with function hommap.mm.2.hs.symbols
}
\keyword{datasets}
